requirements.txt
opentelemetry-exporter-otlp-proto-http==1.17.0
opentelemetry-distro==0.38b0
opentelemetry-instrumentation-aws-lambda==0.38b0